requirements.txt
python-constraint
requests
pygame